requirements.txt
flask
python-dotenv
google-genai 
werkzeug==2.3.7
